Implemented MVC architecture for Registration.
